reverted back to db check for permissions so its always up to date
