+++ server/util/session.js
@@ -1,40 +1,15 @@
-const config = require('./../config');
+const { randomString } = require('./helpers');
+
 const SESSION_COOKIE_NAME = 'session';
-const { sendSuccess } = require('./helpers')
 
-function createSession(res, user) {
-    let str = JSON.stringify(user);
-    let encoded = encodeWithSecret(str);
-    user.session_cookie = encoded;
-    res.cookie(SESSION_COOKIE_NAME, encoded, { httpOnly: true });
-    return sendSuccess(res, user);
-}
-
-function encodeWithSecret(str) {
-    if (!str) return str;
-    let newStr = '';
-    for (let i = 0, j = 0; i < str.length; i++ , j++) {
-        if (j > config.secret.length - 1) j = 0;
-        let encodedCharCode = str.charCodeAt(i) + config.secret.charCodeAt(j);
-        newStr += String.fromCharCode(encodedCharCode);
-    }
-    return newStr;
-}
-
-function decodeWithSecret(str) {
-    if (!str) return str;
-    let newStr = '';
-    for (let i = 0, j = 0; i < str.length; i++ , j++) {
-        if (j > config.secret.length - 1) j = 0;
-        let decodedCharCode = str.charCodeAt(i) - config.secret.charCodeAt(j);
-        newStr += String.fromCharCode(decodedCharCode);
-    }
-    return newStr;
+function createSession(req, res, user) {
+    let db = req.app.get('db');
+    let sessionCookie = randomString(30);
+    res.cookie(SESSION_COOKIE_NAME, sessionCookie, { httpOnly: true });
+    return db.users.update({ id: user.id }, { session_cookie: sessionCookie })
 }
 
 module.exports = {
     SESSION_COOKIE_NAME,
     createSession,
-    encodeWithSecret,
-    decodeWithSecret
 }
